dummy_app/requirements.txt
python-dateutil>=2.8.0,<3
GitPython==3.1.40
